Restucture graphql directory
